requirements.txt
numpy==1.23.1
opencv-python==4.6.0.66
python-dotenv==0.20.0
